+++ source/redux/preload.js
@@ -7,7 +7,7 @@ import hoist_statics from 'hoist-non-react-statics'
     Otherwise, we cannot find and call the preload and preload_deferred methods.
 */
 
-export default function(preload, preload_deferred)
+export default function(parameters)
 {
 	return function(Wrapped)
 	{
@@ -19,7 +19,23 @@ export default function(preload, preload_deferred)
 			}
 		}
 
+		let preload
+		let preload_deferred
+		let preload_blocking
+
+		if (typeof parameters === 'function')
+		{
+			preload = parameters
+		}
+		else
+		{
+			preload          = parameters.preload
+			preload_deferred = parameters.deferred
+			preload_blocking = parameters.blocking
+		}
+
 		Wrapper.preload = preload
+		Wrapper.preload_blocking = preload_blocking
 		Wrapper.preload_deferred = preload_deferred
 
 		return hoist_statics(Wrapper, Wrapped)